+++ Magic/CHANGELOG.md
@@ -6,6 +6,7 @@
  - Lightsabers converted to be class-based
  - Blaster skill upgrades with use
  - Bending powers converted to be class-based and work like the spell book
+ - Added support for SkillAPI attributes in spell configs
 
 ## 7.0